move fwjr server into the common server, but fwjr has ver different data structure from the cmspop and xrootd, so we have to keep the data parsing.
